feat(ssoma): proyectos habilitados, ajustes horas hombre/indicadores/PASO salud
- Nueva feature ProyectoHabilitadoFeature: habilitación de proyectos por
  módulo SSOMA (tabla ss_proyecto_habilitado), independiente del active/estado
  genérico de project, para que cada módulo cure su propio subconjunto.
- Migración AddSsProyectoHabilitado.
- Ajustes en Horas Hombre e Indicadores Proactivos (dtos/repositorios).
- Reescritura de GetActividadesSaludAsync en PasoService_FIXED para expandir
  a todos los proyectos/meses en vez de solo ejecuciones ya generadas.
- Nuevo HorarioLaboralCalculator en Shared.

+++ Features/SsomaModule/IndicadoresProactivosFeature/Application/Dtos/IndicadoresReactivosDtos.cs
@@ -10,7 +10,7 @@ public record IndicadorReactivoProyectoDto
     public int Anio { get; init; }
 
     // Horas Hombre Trabajadas del período
-    public long HorasHombreTrabajadas { get; init; }
+    public decimal HorasHombreTrabajadas { get; init; }
 
     // Contadores base
     public int TotalAccidentes { get; init; }
